Refactor portfolio data imports and polish header and card styles
