+++ src/components/portfolio/portfolio.css
@@ -18,7 +18,7 @@ h6 {
   border: 1px solid transparent;
   transition: var(--transition);
   backdrop-filter: blur(15px);
-  height: 400px; /* fixed height */
+  height: 500px; /* fixed height */
   display: flex; /* display as flex container */
   flex-direction: column; /* flex items should be stacked vertically */
 }
@@ -31,7 +31,14 @@ h6 {
 .portfolio__item-image {
   border-radius: 1.5rem;
   overflow: hidden;
-  height: 30%;
+  height: 35%;
+}
+
+.portfolio__item-image img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  display: block;
 }
 
 .portfolio__item h3 {
